Fixed bug in free chlorine giardia rounding, and fixed bug in temp col for free chlorine giardia

+++ database/migrations/2017_09_26_171824_create_free_chlorine_giardia_inactivations_table.php
@@ -15,7 +15,7 @@ class CreateFreeChlorineGiardiaInactivationsTable extends Migration
     {
         Schema::create('free_chlorine_giardia_inactivations', function (Blueprint $table) {
           $table->increments('id');
-          $table->unsignedTinyInteger('temperature');
+          $table->float('temperature', 3, 1);
           $table->float('ph', 2, 1);
           $table->float('log_inactivation', 2, 1);
           $table->float('disinfectant_concentration', 2, 1);
